api_library/slack☑/requirements.txt
python-dotenv
slack-sdk